requirements.txt
chainlit
python-dotenv
openai